backend/pyproject.toml
[project]
name = "wrapiq-backend"
version = "0.1.0"
description = "WrapIQ Backend API"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "psycopg[binary]>=3.2.0",
    "alembic>=1.14.0",
    "pydantic[email]>=2.9.0",
    "pydantic-settings>=2.6.0",
    "redis[hiredis]>=5.2.0",
    "arq>=0.26.0",
    "httpx>=0.28.0",
    "pyjwt>=2.9.0",
    "bcrypt>=4.2.0",
    "resend>=2.5.0",
    "python-multipart>=0.0.9",
    "google-genai>=1.0.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "mypy>=1.13.0",
]

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
